+++ lib/pageData/[...paths].ts
@@ -3,7 +3,7 @@ import { getFilterSchema, pathArrayToFilterArray } from '../filters'
 import fs from 'fs'
 import path from 'path'
 import matter from 'gray-matter'
-import getChilliData from '~/lib/getChilliData'
+import { getChilliData, getSingleChilli } from '~/lib/chilliData'
 
 export const getChilliPageDataFromPaths = async (paths: string[]): Promise<ICultivarPageData> => {
   let chillies: ICultivar[] = []
@@ -47,46 +47,13 @@ export const getChilliPageDataFromPaths = async (paths: string[]): Promise<ICult
 
       chillies = await getChilliData()
     } else if (paths[0] === 'cultivars') {
-      console.log('this is a handle page')
       //this is a handle page
-      // requestType = 'handle'
+      requestType = 'handle'
       // const filterFormula = `{handle}="${paths[1]}"`
-      const data = await getChilliData()
-      chillies = data
-
-      //try and get at least 4 related chillies
-      // const chilli = data[0]
-      // if (chilli) {
-      //   const min = 4
-
-      //   //same origin AND species
-      //   relatedChillies.push(
-      //     ...(await getChilliesFromAirtable({
-      //       filterFormula: `AND(FIND("${chilli.species?.handle}", {species/handle}), FIND("${chilli.origin?.handle}", {origin/handle}), NOT({handle}="${chilli.handle}"))`,
-      //     }))
-      //   )
-
-      //   //same origin
-      //   if (relatedChillies.length < min) {
-      //     relatedChillies.push(
-      //       ...(await getChilliesFromAirtable({
-      //         filterFormula: `AND(FIND("${chilli.origin?.handle}", {origin/handle}), NOT({handle}="${chilli.handle}"))`,
-      //       }))
-      //     )
-      //   }
-
-      //   //same species
-      //   if (relatedChillies.length < min) {
-      //     relatedChillies.push(
-      //       ...(await getChilliesFromAirtable({
-      //         filterFormula: `AND(FIND("${chilli.species?.handle}", {species/handle}), NOT({handle}="${chilli.handle}"))`,
-      //       }))
-      //     )
-      //   }
-
-      //   relatedChillies = [...new Map(relatedChillies.map((item) => [item.handle, item])).values()] // unique
-      //   relatedChillies = shuffle(relatedChillies).slice(0, 4) // get random 4
-      // }
+      const handle = paths[1]
+      if (handle) {
+        chillies = [await getSingleChilli(handle)]
+      }
     } else if (paths.length > 1) {
       //   //do we have a sort by?
       //   const last = paths[paths.length - 1]
@@ -105,8 +72,6 @@ export const getChilliPageDataFromPaths = async (paths: string[]): Promise<ICult
       const filterPaths = chunk(hasSort ? paths.slice(0, -1) : paths)
 
       filters = pathArrayToFilterArray(filterPaths, schema)
-
-      console.log({ filters })
 
       if (filters) {
         requestType = 'listing'
